Adds "validation" and "debug_warnings" features.
This code got a little gnarley, but this now enables capability to
disable the debug warnings (which are exclusive to debug builds)
as well as validation (the idea being that you've tested your
inputs and don't want to execute validation code for values you
are already certain to be valid.  The default is *still* to use
validation as well as debug warnings, but you can go "bare-metal"
with `--no-default-features` treatment in your Cargo.toml's
dependencies stanza.