Fixes all flavors of build and feature.
Split up validation and warning functionality into two traits.
Thought there was that trait default functions could provide
"no-op" implementations when features were not in use, but it
was just as easy to annotate feature usage on the method calls
(so I removed the default methods). Code looked good as far as
distinguishing its responsibility, so I left it as trait impl
for the enum rather than adding an impl on the enum.. 6 and 1/2
dozen, as they say.