Validation warnings feature (#1)
* Adds "validation" and "debug_warnings" features.

This code got a little gnarley, but this now enables capability to
disable the debug warnings (which are exclusive to debug builds)
as well as validation (the idea being that you've tested your
inputs and don't want to execute validation code for values you
are already certain to be valid.  The default is *still* to use
validation as well as debug warnings, but you can go "bare-metal"
with `--no-default-features` treatment in your Cargo.toml's
dependencies stanza.

* Fixes all flavors of build and feature.

Split up validation and warning functionality into two traits.
Thought there was that trait default functions could provide
"no-op" implementations when features were not in use, but it
was just as easy to annotate feature usage on the method calls
(so I removed the default methods). Code looked good as far as
distinguishing its responsibility, so I left it as trait impl
for the enum rather than adding an impl on the enum.. 6 and 1/2
dozen, as they say.

---------

Co-authored-by: Andrew Oswald <aoswald@chariotsolutions.com>